Reorganize mutation code.
